+++ packages/central-server/__tests__/sync/CentralSyncManager.setupSnapshotForPull.test.js
@@ -17,7 +17,12 @@ import {
 } from '@tamanu/constants';
 import { toDateTimeString } from '@tamanu/utils/dateTime';
 
-import { createTestContext, waitForSession, waitForPushCompleted } from '../utilities';
+import {
+  createTestContext,
+  waitForSession,
+  waitForPushCompleted,
+  initializeCentralSyncManagerWithContext,
+} from '../utilities';
 import { importerTransaction } from '../../dist/admin/importer/importerEndpoint';
 import { referenceDataImporter } from '../../dist/admin/referenceDataImporter';
 
@@ -37,25 +42,8 @@ describe('CentralSyncManager.setupSnapshotForPull', () => {
   let sequelize;
 
   const DEFAULT_MAX_RECORDS_PER_SNAPSHOT_CHUNKS = 100000000;
-  const DEFAULT_CONFIG = {
-    sync: {
-      lookupTable: {
-        enabled: false,
-      },
-      maxRecordsPerSnapshotChunk: 1000000000,
-    },
-  };
-
-  const initializeCentralSyncManager = config => {
-    // Have to load test function within test scope so that we can mock dependencies per test case
-    const {
-      CentralSyncManager: TestCentralSyncManager,
-    } = require('../../dist/sync/CentralSyncManager');
-
-    TestCentralSyncManager.overrideConfig(config || DEFAULT_CONFIG);
-
-    return new TestCentralSyncManager(ctx);
-  };
+  const initializeCentralSyncManager = config =>
+    initializeCentralSyncManagerWithContext(ctx, config);
 
   const prepareRecordsForSync = async () => {
     // Pre insert the records below for snapshotting later
