+++ packages/central-server/__tests__/sync/CentralSyncManagerEdgeCases.test.js
@@ -4,32 +4,20 @@ import { fake } from '@tamanu/fake-data/fake';
 import { APPOINTMENT_STATUSES, REPEAT_FREQUENCY, SYSTEM_USER_UUID } from '@tamanu/constants';
 import { settingsCache } from '@tamanu/settings';
 
-import { createTestContext, waitForSession, waitForPushCompleted } from '../utilities';
+import {
+  createTestContext,
+  waitForSession,
+  waitForPushCompleted,
+  initializeCentralSyncManagerWithContext,
+} from '../utilities';
 
 describe('CentralSyncManager Edge Cases', () => {
   let ctx;
   let models;
 
   const DEFAULT_MAX_RECORDS_PER_SNAPSHOT_CHUNKS = 100000000;
-  const DEFAULT_CONFIG = {
-    sync: {
-      lookupTable: {
-        enabled: false,
-      },
-      maxRecordsPerSnapshotChunk: 1000000000,
-    },
-  };
-
-  const initializeCentralSyncManager = config => {
-    // Have to load test function within test scope so that we can mock dependencies per test case
-    const {
-      CentralSyncManager: TestCentralSyncManager,
-    } = require('../../dist/sync/CentralSyncManager');
-
-    TestCentralSyncManager.overrideConfig(config || DEFAULT_CONFIG);
-
-    return new TestCentralSyncManager(ctx);
-  };
+  const initializeCentralSyncManager = config =>
+    initializeCentralSyncManagerWithContext(ctx, config);
 
   beforeAll(async () => {
     ctx = await createTestContext();
