+++ packages/central-server/__tests__/sync/sensitiveFacilities.test.js
@@ -3,32 +3,19 @@ import { fake } from '@tamanu/fake-data/fake';
 import { SETTINGS_SCOPES, SYSTEM_USER_UUID } from '@tamanu/constants';
 import { getCurrentDateTimeString } from '@tamanu/utils/dateTime';
 
-import { createTestContext, waitForSession } from '../utilities';
+import {
+  createTestContext,
+  waitForSession,
+  initializeCentralSyncManagerWithContext,
+} from '../utilities';
 
 describe('CentralSyncManager', () => {
   let ctx;
   let models;
 
   const DEFAULT_MAX_RECORDS_PER_SNAPSHOT_CHUNKS = 100000000;
-  const DEFAULT_CONFIG = {
-    sync: {
-      lookupTable: {
-        enabled: false,
-      },
-      maxRecordsPerSnapshotChunk: 1000000000,
-    },
-  };
-
-  const initializeCentralSyncManager = config => {
-    // Have to load test function within test scope so that we can mock dependencies per test case
-    const {
-      CentralSyncManager: TestCentralSyncManager,
-    } = require('../../dist/sync/CentralSyncManager');
-
-    TestCentralSyncManager.overrideConfig(config || DEFAULT_CONFIG);
-
-    return new TestCentralSyncManager(ctx);
-  };
+  const initializeCentralSyncManager = config =>
+    initializeCentralSyncManagerWithContext(ctx, config);
 
   beforeAll(async () => {
     ctx = await createTestContext();
